LPS-136108 Create a new test class specifically to test the LayoutReferencesExportImportContentProcessor class.
While we previously used the DefaultExportImportContentProcessorTest for
this purpose, I believe it would be beneficial to create this new test
class so we can create finer, more detailed tests that test all the many
edge cases in the LayoutReferencesExportImportContentProcessor class.

In the DefaultExportImportContentProcessorTest, the tests we run are at
more of a "high" level. We run the content processor on a file that
contains a large number of URLs, but we don't have a good mechanism to
check the exact values of the resulting URLs. All we can do is check
that the file contains or doesn't contain certain strings.

In this new LayoutReferencesExportImportContentProcessor test class, we
will create tests that consist of setting up Groups and Layouts in
various ways, and then testing a single URL per test, to make sure that
the resulting imported URL is as expected.

In this first commit, we will just add one simple proof-of-concept test
to test that a non-Liferay URL retains it's same values. We will add
many more test cases in the next few commits.